requirements.txt
AWSIoTPythonSDK==1.5.4
blinker==1.8.2
boto3==1.34.113
botocore==1.34.113
click==8.1.7
colorama==0.4.6
docutils==0.16
Flask==3.0.3
future==1.0.0
iso8601==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
MarkupSafe==2.1.5
numpy==1.26.4
paho-mqtt==2.1.0
pyasn1==0.6.0
pyserial==3.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.1
rsa==4.7.2
s3transfer==0.10.1
serial==0.0.97
six==1.16.0
urllib3==2.2.1
Werkzeug==3.0.3